+++ packages/types/custom/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## unpublished
 
+- Feature: Add new parameter `measureOptions` to `DrawConfiguration`.
+- Feature: Add new type `MeasureOptions`.
+- Feature: Add new type `MeasureMode`.
+- Feature: Add optional property `measure` to `DrawStyle`.
 - Fix: Make `selectionStyle`, `hoverStyle`, `defaultStyle` and `unselectableStyle` optional and edit type for `dispatchOnMapSelect` in interface `ExtendedMasterportalapiMarkers`. 
 
 ## 1.5.0